fix: remove filter sidebar to separate file and add ignore in config
